restructuring the project with folders and creating a new screen to save BMI results
