+++ lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'home_page.dart';
-import 'second_page.dart';
+import 'screens/home_page.dart';
+import 'screens/second_page.dart';
 
 void main() => runApp(MyApp());
