+++ lib/screens/saved_bmi_page.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+
+class SavedBMICalculation extends StatefulWidget {
+  @override
+  _SavedBMICalculationState createState() => _SavedBMICalculationState();
+}
+
+class _SavedBMICalculationState extends State<SavedBMICalculation> {
+  @override
+  Widget build(BuildContext context) {
+    return Container();
+  }
+}
